implement Fileinfo for Windows and non-Windows
